Moved worker code to it's own file
