+++ src/index.js
@@ -3,6 +3,7 @@ import { Client, GatewayIntentBits } from 'discord.js';
 import pino from 'pino';
 import instances from './model/instances.js';
 import { useCommands } from './bot/commands/index.js';
+import { useWorker } from './bot/worker.js';
 
 const logger = pino();
 const client = new Client({
@@ -14,80 +15,11 @@ const client = new Client({
   ],
 });
 
-instances.fetch();
-useCommands(client);
-
 client.on('ready', async () => {
-  logger.info(`${client.user.tag} is online`);
-  setInterval(async () => {
-    const startTime = Date.now();
-    for (const instance of instances.cache) {
-      logger.info(instance, 'running instance');
-      const channel = await _getChannelFromInstance(instance);
-      if (!channel) break;
+  await instances.fetch();
+  await useCommands(client);
+  useWorker(client);
 
-      const messages = await channel.messages.fetch({ limit: 100 });
-      logger.info({ count: messages.size }, 'fetched messages');
-      const cutoff = Date.now() - instance.max_message_age;
-
-      for (const [id, message] of messages) {
-        // Ignore messages with the :floppy_disk: reaction.
-        if (await _hasReaction(message, '💾')) break;
-
-        // Ignore messages that are young.
-        if (message.createdTimestamp > cutoff) break;
-
-        // Delete the rest of the messages.
-        try {
-          await message.delete();
-          logger.info({ id: message.id }, 'deleted message');
-        } catch (err) {
-          if (err.status == 404) {
-            logger.warn(err);
-          } else {
-            logger.error(err);
-          }
-        }
-      }
-    }
-
-    logger.info({ duration: Date.now() - startTime }, 'completed run');
-  }, 10000);
+  logger.info(`${client.user.tag} is online!`);
 });
-/**
- * Gets a channel or undefined if something went wrong.
- * @param {import('./instance.js').Instance} instance
- * @returns {Promise<import('discord.js').TextChannel|null|undefined>}
- */
-async function _getChannelFromInstance(instance) {
-  try {
-    const channel = await client.channels.fetch(instance.channel_id);
-    if (!channel.isTextBased()) {
-      logger.error('instance was running on non-text channel');
-      await repo.delete(instance.id);
-      return undefined;
-    }
-    return channel;
-  } catch (err) {
-    logger.error(err, 'failed to fetch channel');
-    await instances.delete(instance.id);
-    return undefined;
-  }
-}
-/**
- * Gets whether or not a message has a reaction.
- *
- * @param {Message} message
- * @param {string} key
- * @return {Promise<boolean>}
- */
-async function _hasReaction(message, key) {
-  const reaction = message.reactions.cache.get(key);
-  try {
-    return reaction && (await reaction.users.fetch()).size > 0;
-  } catch {
-    return false;
-  }
-}
-
 client.login(process.env.DISCORD_CLIENT_TOKEN);
